update all bar charts to import options from chart options file
