+++ src/components/CodeQuality.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import chartOptions from '../helpers/ChartOptions';
 
@@ -30,44 +29,12 @@ class CodeQuality extends React.Component {
       ]
     };
 
-    const options = {
-      fullWidth: true,
-      chartPadding: {
-        left: 10,
-        top: 20,
-        right: 10,
-        bottom: 10,
-      },
-      height: 300,
-      high: 50,
-      low: 0,
-      axisY: {
-        onlyInteger: true,
-      },
-      plugins: [
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-        Chartist.plugins.ctAxisTitle({
-          axisY: {
-            axisTitle: 'Percent',
-            axisClass: 'ct-axis-title',
-            textAnchor: 'middle',
-            flipTitle: false,
-          }
-        })
-      ]
-    };
-
     return (
       <div>
         <ChartistGraph
           data={data}
+          options={chartOptions.barOptions}
           responsiveOptions={chartOptions.barResponsiveOptions}
-          options={options}
           type={'Bar'} />
       </div>
