+++ src/components/NotUpdated.js
@@ -1,8 +1,5 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
-import 'chartist-plugin-tooltips';
-import 'chartist-plugin-axistitle';
 import chartOptions from '../helpers/ChartOptions';
 
 class NotUpdated extends React.Component {
@@ -29,45 +26,11 @@ class NotUpdated extends React.Component {
       ]
     };
 
-
-    const options = {
-      fullWidth: true,
-      chartPadding: {
-        left: 10,
-        top: 20,
-        right: 10,
-        bottom: 10,
-      },
-      height: 300,
-      high: 50,
-      low: 0,
-      axisY: {
-        onlyInteger: true,
-      },
-      plugins: [
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-        Chartist.plugins.ctAxisTitle({
-          axisY: {
-            axisTitle: 'Percent',
-            axisClass: 'ct-axis-title',
-            textAnchor: 'middle',
-            flipTitle: false,
-          }
-        })
-      ]
-    };
-
-
     return (
       <div>
         <ChartistGraph
           data={data}
-          options={options}
+          options={chartOptions.barOptions}
           responsiveOptions={chartOptions.barResponsiveOptions}
           type={'Bar'} />
       </div>
