+++ src/components/RailsVersions.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import chartOptions from '../helpers/ChartOptions';
 
@@ -37,42 +36,11 @@ class RailsVersions extends React.Component {
       ]
     };
 
-    const options = {
-      fullWidth: true,
-      chartPadding: {
-        left: 10,
-        top: 20,
-        right: 10,
-      },
-      height: 300,
-      high: 30,
-      low: 0,
-      axisY: {
-        onlyInteger: true,
-      },
-      plugins: [
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-        Chartist.plugins.ctAxisTitle({
-          axisY: {
-            axisTitle: 'Percent',
-            axisClass: 'ct-axis-title',
-            textAnchor: 'middle',
-            flipTitle: false,
-          }
-        })
-      ]
-    };
-
     return (
       <div>
         <ChartistGraph
           data={data}
-          options={options}
+          options={chartOptions.barOptions}
           responsiveOptions={chartOptions.barResponsiveOptions}
           type={'Bar'} />
       </div>
